+++ tests/fixtures/migrations/alter_field_without_default_make_nullable_with_db_default.py
@@ -1,0 +1,18 @@
+from django.db import migrations, models
+
+
+class Migration(migrations.Migration):
+    app_label = "mock_app"
+    name = "remove_field"
+
+    dependencies = [
+        ("testapp", "0002_add_email"),
+    ]
+
+    operations = [
+        migrations.AlterField(
+            model_name="user",
+            name="non_null_field_without_default",
+            field=models.CharField(null=True, db_default="something"),
+        ),
+    ]